// src/renderer/src/app/enums/settings-section.ts
/**
 * @license
 * Copyright (c) Noldova.
 *
 * This source code is licensed under the license found in the
 * LICENSE file in the root directory of this source tree.
 */

export enum SettingsSection {
  General = "General",
  Appearance = "Appearance",
  Providers = "Providers",
  Teammates = "Teammates",
  Shortcuts = "Shortcuts",
  Gallery = "Gallery",
  About = "About"
}